clean duplicate and unused code
