+++ routes/video_utils.py
@@ -1,0 +1,77 @@
+import os
+import shutil
+import boto3
+from typing import Union
+from dotenv import load_dotenv
+
+# Load environment variables
+load_dotenv()
+
+# Configuration
+USE_LOCAL_STORAGE = os.getenv("USE_LOCAL_STORAGE") or False
+BASE_URL = os.getenv("BASE_URL")
+DO_SPACES_ACCESS_KEY = os.getenv("DO_SPACES_ACCESS_KEY")
+DO_SPACES_ACCESS_SECRET = os.getenv("DO_SPACES_ACCESS_SECRET")
+DO_SPACES_REGION = os.getenv("DO_SPACES_REGION")
+DO_SPACES_BUCKET = os.getenv("DO_SPACES_BUCKET")
+DO_SPACES_ENDPOINT = os.getenv("DO_SPACES_ENDPOINT")
+
+
+def upload_to_digital_ocean_storage(file_path: str, video_storage_file_name: str) -> str:
+    """
+    Uploads the video to DigitalOcean Spaces and returns the public URL.
+    """
+    client = boto3.client(
+        's3',
+        region_name=DO_SPACES_REGION,
+        endpoint_url=DO_SPACES_ENDPOINT,
+        aws_access_key_id=DO_SPACES_ACCESS_KEY,
+        aws_secret_access_key=DO_SPACES_ACCESS_SECRET
+    )
+
+    # Ensure the file name has .mp4 extension
+    if not video_storage_file_name.endswith('.mp4'):
+        video_storage_file_name += '.mp4'
+
+    client.upload_file(
+        file_path,
+        DO_SPACES_BUCKET,
+        video_storage_file_name,
+        ExtraArgs={'ACL': 'public-read', 'ContentType': 'video/mp4'}
+    )
+
+    return f"{DO_SPACES_ENDPOINT}/{DO_SPACES_BUCKET}/{video_storage_file_name}"
+
+
+def move_to_public_folder(
+    file_path: str, video_storage_file_name: str, base_url: Union[str, None] = None
+) -> str:
+    """
+    Moves the video to the public folder and returns the URL.
+    """
+    public_folder = os.path.join(os.path.dirname(__file__), "public")
+    os.makedirs(public_folder, exist_ok=True)
+
+    new_file_name = f"{video_storage_file_name}.mp4"
+    new_file_path = os.path.join(public_folder, new_file_name)
+
+    shutil.move(file_path, new_file_path)
+
+    # Use the provided base_url if available, otherwise fall back to BASE_URL
+    url_base = base_url if base_url else BASE_URL
+    video_url = f"{url_base.rstrip('/')}/public/{new_file_name}"
+    return video_url
+
+
+def get_frame_config(aspect_ratio):
+    """
+    Returns frame configuration based on aspect ratio.
+    """
+    if aspect_ratio == "16:9":
+        return (3840, 2160), 14.22
+    elif aspect_ratio == "9:16":
+        return (1080, 1920), 8.0
+    elif aspect_ratio == "1:1":
+        return (1080, 1080), 8.0
+    else:
+        return (3840, 2160), 14.22
